pyproject.toml
[project]
name = "graphons"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "einops>=0.8.1",
    "guacamol>=0.5.2",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pot>=0.9.6.post1",
    "rdkit>=2025.9.3",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.8.0",
    "torch~=2.8.0",
    "torch-geometric-temporal>=0.56.2",
    "torchvision>=0.23.0",
    "umap-learn>=0.5.9.post2",
    "wandb>=0.23.1",
    "xgboost>=3.1.3",
]
